+++ assets/js/console.js
@@ -1,36 +1,48 @@
 // This script will be run within the webview itself
 // It cannot access the main VS Code APIs directly.
-(function () {
+const vscode = acquireVsCodeApi();
 
-  const vscode = acquireVsCodeApi();
+// const oldState = vscode.getState();
+// console.log(oldState);
+// vscode.setState({ count: currentCount });
 
-  // const oldState = vscode.getState();
-  // console.log(oldState);
-  // vscode.setState({ count: currentCount });
+function jumpToPageBottom() {
+  $('html, body').scrollTop( $(document).height()-$(window).height());
+}
 
-  function jumpToPageBottom() {
-    $('html, body').scrollTop( $(document).height()-$(window).height());
-  }
+function isAtBottom() {
+  return $(window).scrollTop() + $(window).height() === $(document).height();
+}
 
-  function isAtBottom() {
-    return $(window).scrollTop() + $(window).height() === $(document).height();
-  }
+// Handle messages sent from the extension to the webview
+window.addEventListener('message', event => {
+  var wasAtBottom = isAtBottom()
+  $("<div />").append($.parseHTML(event.data)).appendTo("#data")
+  if (wasAtBottom) jumpToPageBottom()
+})
 
-  // // Send a message back to the extension
-  // vscode.postMessage({
-  //   command: 'alert',
-  //   text: '🐛  on line ' + currentCount
-  // });
+$(document).dblclick(function() {
+  $('#commandInput>input').focus()
+})
 
-  // Handle messages sent from the extension to the webview
-  window.addEventListener('message', event => {
-    const message = event.data; // The json data that the extension sent
-    switch (message.command) {
-        case 'append':
-            var wasAtBottom = isAtBottom()
-            $("<div />").text("Message: " + message.text).appendTo("#data")
-            if (wasAtBottom) jumpToPageBottom()
-            break;
+$(document).bind('mousewheel', function (event) {
+  // do your stuff
+  if (event.ctrlKey) {
+    if (event.originalEvent.wheelDelta > 0 || event.originalEvent.detail < 0) {
+      // TODO: Font Size up
+    } else {
+      // TODO: Font Size down
     }
-  })
-}())
+  }
+});
+
+$("#commandInput>input").on('keyup', function (e) {
+  if (e.keyCode == 13) {
+      // Send a message back to the extension
+      vscode.postMessage({
+        command: 'customMessage',
+        text: $("#commandInput>input").val().substring(1)
+      })
+      $("#commandInput>input").val(">")
+  }
+});